chore(travis): run tests on Travis-CI
PR #10 generously added a nice test suite. It might change in the future (admittedly,
probably not), but it's certainly a good start, and running tests on Travis will help
me be confident in merging patches.

Of course it's not at all an exhaustive suite, this should be improved over time. But
it's not bad.

Closes #13