fix(bower-release): don't extend dependencies from source repository unless requested
It is now possible to request to extend the source repository's dependencies, using the options
key `extendDependencies` with the value `true`.

This is not yet reflected in documentation. Unfortunately, the infrastructure of this project
is such that writing a test for this is very difficult, so I'm just going to have to hope that
it works.

I do hope to clean this up down the line, but you know how time is, fleeting.

Closes #31
Closes #32